Seperated the eventbrite service responsibilities into the client and the caching of events and renamed appropriatly
