+++ frontend/app/Global.scala
@@ -4,15 +4,15 @@ import play.api.Application
 import play.api.mvc.WithFilters
 import play.filters.csrf._
 import services._
-import services.eventbrite.{GuardianLiveEventService, MasterclassEventService, LocalEventService}
+import services.eventbrite.{GuardianLiveEventCache, MasterclassEventCache, LocalEventCache}
 
 object Global extends WithFilters(RedirectMembersFilter, CheckCacheHeadersFilter, CSRFFilter(), Gzipper, AddEC2InstanceHeader) {
   override def onStart(app: Application) {
     SentryLogging.init()
 
-    GuardianLiveEventService.start()
-    LocalEventService.start()
-    MasterclassEventService.start()
+    GuardianLiveEventCache.start()
+    LocalEventCache.start()
+    MasterclassEventCache.start()
 
     GuardianContentService.start()
   }
