+++ frontend/app/controllers/WhatsOn.scala
@@ -8,14 +8,14 @@ import model.RichEvent._
 import model._
 import play.api.mvc.Controller
 import services._
-import services.eventbrite.{GuardianLiveEventService, MasterclassEventService, LocalEventService, EventbriteService}
+import services.eventbrite._
 import tracking.ActivityTracking
 
 trait WhatsOn extends Controller with ActivityTracking {
 
-  val guLiveEvents: EventbriteService
-  val localEvents: EventbriteService
-  val masterclassEvents: EventbriteService
+  val guLiveEvents: EventbriteCache
+  val localEvents: EventbriteCache
+  val masterclassEvents: EventbriteCache
 
   private def allEvents = {
     guLiveEvents.events ++ localEvents.events
@@ -101,7 +101,7 @@ trait WhatsOn extends Controller with ActivityTracking {
 }
 
 object WhatsOn extends WhatsOn {
-  val guLiveEvents = GuardianLiveEventService
-  val localEvents = LocalEventService
-  val masterclassEvents = MasterclassEventService
+  val guLiveEvents = GuardianLiveEventCache
+  val localEvents = LocalEventCache
+  val masterclassEvents = MasterclassEventCache
 }
