+++ frontend/app/services/eventbrite/GuardianLiveEventCache.scala
@@ -13,8 +13,7 @@ import scala.concurrent.duration._
 import play.api.Play.current
 
 
-object GuardianLiveEventService extends LiveService {
-  val apiToken = Config.eventbriteApiToken
+object GuardianLiveEventCache extends LiveEventCache {
   // For partner/patrons with free event tickets benefits, we generate a discount code which unlocks a combination of
   // maximum 2 discounted tickets and 1 complimentary ticket.
   // The maxDiscountQuantityAvailable value is used to set the Access code 'quantity_available' attribute (i.e. the
@@ -22,9 +21,9 @@ object GuardianLiveEventService extends LiveService {
   //
   // see https://www.eventbrite.com/developer/v3/formats/event/#ebapi-access-code
   val maxDiscountQuantityAvailable = 3
-  val wsMetrics = new EventbriteMetrics("Guardian Live")
-
+  val client = new EventbriteClient(Config.eventbriteApiToken, maxDiscountQuantityAvailable, new EventbriteMetrics("Guardian Live"))
   val refreshTimePriorityEvents = new FiniteDuration(Config.eventbriteRefreshTimeForPriorityEvents, SECONDS)
+
   lazy val eventsOrderingTask = ScheduledTask[Seq[String]]("Event ordering", Nil, 1.second, refreshTimePriorityEvents) {
     for {
       ordering <- WS.url(Config.eventOrderingJsonUrl).get()
@@ -34,7 +33,7 @@ object GuardianLiveEventService extends LiveService {
   def mkRichEvent(event: EBEvent): Future[RichEvent] = for { gridImageOpt <- gridImageFor(event) }
     yield GuLiveEvent(event, gridImageOpt, contentApiService.content(event.id))
 
-  override def getFeaturedEvents: Seq[RichEvent] = EventbriteServiceHelpers.getFeaturedEvents(eventsOrderingTask.get(), events)
+  override def getFeaturedEvents: Seq[RichEvent] = EventbriteCacheHelpers.getFeaturedEvents(eventsOrderingTask.get(), events)
   override def start() {
     super.start()
     eventsOrderingTask.start()
